Fixing tasks context saving
In previous commit the context saving was not done correclty,
hence the tasks were not completing as they should. They were
printing less information.

To fix this issues many changes were performed:
thread_start
- In previous commit, the LDMIA was performed before the SPSEL
was changed, that's an error since we need to perform that operation
using the PSP.

PendSV_Handler
- The most relevant issue before was that task[lastTask] information
was re-load into the scheduler (as it was the first time it's going
to run), instead of saving it's context and re-starting from it the
next time.
- Here the code was changed to store the registers values into memory,
and then that memory address into the C structure tasks[lastTask].
- For that reason a STMDB was used nstead of STMIA. Because we
don't want to miss the address memory holding the registers.
- In previous commit, the instruction was moving the tasks[lastTask]
into the registers, and never modifying that structure.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>